refactor: router setting
