+++ package.json
@@ -16,7 +16,8 @@
     "@mui/material": "^5.15.10",
     "firebase": "^10.8.0",
     "react": "^18.2.0",
-    "react-dom": "^18.2.0"
+    "react-dom": "^18.2.0",
+    "react-router-dom": "^6.22.1"
   },
   "devDependencies": {
     "@types/react": "^18.2.55",
