+++ src/App.tsx
@@ -1,12 +1,17 @@
-import AddTodo from "./presentation/AddTodo"
-import GetTodo from "./presentation/GetTodo"
+import { Routes, Route } from 'react-router-dom';
+import AddTodo from "./presentation/component/AddTodo"
+import GetTodo from "./presentation/component/GetTodo"
+import NoMatch from './presentation/page/NoMatch';
 
 function App() {
 
   return (
     <div>
-      <AddTodo />
-      <GetTodo />
+      <Routes>
+        <Route path="/" element={<GetTodo />} />
+        <Route path="/add" element={<AddTodo />} />
+        <Route path="*" element={<NoMatch />} />
+      </Routes>
     </div>
   )
 }
