+++ src/presentation/page/AddTodoPage.tsx
@@ -1,0 +1,12 @@
+import AddTodo from "../component/AddTodo";
+
+const AddTodoPage: React.FC = () => {
+  return (
+    <div>
+      <h1>やることを追加</h1>
+      <AddTodo />
+    </div>
+  );
+}
+
+export default AddTodoPage;
