+++ src/presentation/page/GetTodoPage.tsx
@@ -1,0 +1,12 @@
+import GetTodo from "../component/GetTodo";
+
+const GetTodoPage: React.FC = () => {
+    return (
+        <div>
+        <h1>やることを取得</h1>
+        <GetTodo />
+        </div>
+    );
+};
+
+export default GetTodoPage;
